+++ NEWS.md
@@ -1,4 +1,10 @@
 # rSOILWAT2 v6.5.1-devel
+* Simulation output remains the same as the previous version.
+
+## Bugfix
+* `dbW_fixWeather()` now recognizes if the last year is a leap year;
+  this avoids negative precipitation values on the last day
+  which occurred in rare situations (#258; @dschlaep).
 
 
 # rSOILWAT2 v6.5.0